Bump CHANGELOG development version to 0.13.0-dev
This is only an update to the development version and does not represent
a stable release.

The original bump to 0.13.0-dev was made in 9cf2c89, but did not make
any changes to the changelog. This adds a 0.13.0-dev changelog entry to
master so new changes are properly added to the correct place.